+++ infrastructure/opentofu/modules/providers/outscale/versions.tf
@@ -4,5 +4,9 @@ terraform {
       source  = "outscale/outscale"
       version = ">= 0.12.0"
     }
+    talos = {
+      source  = "siderolabs/talos"
+      version = ">= 0.7.0"
+    }
   }
 }