+++ infrastructure/opentofu/modules/providers/ovh/main.tf
@@ -1,15 +1,25 @@
+# Reserve fixed private IPs for control plane nodes
+resource "openstack_networking_port_v2" "control_plane" {
+  count          = var.control_plane_count
+  name           = "${var.cluster_name}-cp-port-${count.index}"
+  network_id     = var.network_id
+  admin_state_up = "true"
+  security_group_ids = [openstack_networking_secgroup_v2.this.id]
+}
+
 resource "openstack_compute_instance_v2" "control_plane" {
-  count       = var.control_plane_count
-  name        = "${var.cluster_name}-cp-${count.index}"
-  image_id    = var.image_id
-  flavor_name = var.flavor_name
-  region      = var.region
+  count             = var.control_plane_count
+  name              = "${var.cluster_name}-cp-${count.index}"
+  image_id          = var.image_id
+  flavor_name       = var.flavor_name
+  region            = var.region
+  availability_zone = element(var.availability_zones, count.index)
 
   network {
-    name = var.network_name
+    port = openstack_networking_port_v2.control_plane[count.index].id
   }
 
-  user_data = var.control_plane_config
+  user_data = data.talos_machine_configuration.control_plane.machine_configuration
 
   security_groups = [openstack_networking_secgroup_v2.this.name]
 
@@ -27,7 +37,7 @@ resource "openstack_compute_instance_v2" "worker" {
     name = var.network_name
   }
 
-  user_data = var.worker_config
+  user_data = data.talos_machine_configuration.worker.machine_configuration
 
   security_groups = [openstack_networking_secgroup_v2.this.name]