+++ infrastructure/opentofu/modules/providers/scw/versions.tf
@@ -4,5 +4,9 @@ terraform {
       source  = "scaleway/scaleway"
       version = ">= 2.40"
     }
+    talos = {
+      source  = "siderolabs/talos"
+      version = ">= 0.7.0"
+    }
   }
 }